Fix link to API docs
Looks like the version-free redirect is broken, so I fixed this branch to the version-specific one.

I did it on this branch only because this is the Google hit I keep getting when I search for said API docs, so I want the link to work; this may be needed on others, too. It's fine on the 2.10 branch.