Add procedure for rotating CredHub set leafs
The previous procedure on the page was only applicable to CAs set in
CredHub. However, it is also possible to set leaf certs in CredHub.
Therefore we need a rotation procedure for these certs.

It should be noted that the leaf cert procedure does not use Maestro in
any way. However, there doesn't seem to be a better place in the
existing docs to add this procedure. Additionally, it has a similar
theme to the other procedures; rotating certs stored in CredHub using a
CLI (maestro or credhub).

[#181067913] Certificate rotation procedures appear on the Certificates page and API endpoint

Signed-off-by: Kenneth Lakin <klakin@vmware.com>