Update stale docs for datastore vs ds cluster usage
it is not required and actually bad to keep datastore names
configured when moving to datastore clusters. the resulting
bosh config will look something like:

```
    datastores:
    - main_datastore_1
    - main_datastore_2
    - main_datastore_3
    - clusters:
      - DatastoreCluster: {}
```

this will instruct bosh to migrate only some of the volumes
on a subsequent recreate of the vms since it's trying to balance
the load across the ds.

update suggestion in case of using datastore clusters to use
the available datastore cluster textfields instead of providing
the names of the datastores in the datastore cluster. That is
exactly why those fields were introduced and these parts of the
docs most probably were overlooked.