Update advanced-certificate-rotation.html.md.erb
The removed section is a remnant from a prior procedure (Manual credhub version) that is not necessary in the maestro packaged with ops manager 2.10+. The instructions listed here are incompletely copied from the prior procedure and cause confusion. After discussion with ops-manager R&D it was agreed to remove this section.