Change references to 2.10 -> 3.0
We have also deleted several items that seem like they are no longer
relevant in modern versions of Ops Manager + TAS.

We also have moved or inlined several partials that were previously
stored in the docs-partials repo. This seems to be an artifact of the
incomplete division of OM and TAS documentation. The partials in
question were only in use in the OM docs.

[#183049507] Create PR to update release notes for OM 3.0

Signed-off-by: Brian Upton <bupton@vmware.com>