OM 3.0 Known Issues: vSphere BOSH Director RAM
With vSphere 3.0, Director has an increased memory footprint. We will
most likely mitigate this in an upcoming release, but we'd like to give
a heads-up to our customers. We provide warning and guidance to
calculate the memory increase.

The larger RAM requirements seems to be caused by a 64x increase in the
resident set size (RSS) of each Ruby thread (16 kiB → 1 MiB). We haven't
isolated the source of the increase, but it seems either
compiler-related or standard library-related.