Merge SSH security group into VMs security group
Previously, the SSH load balancer was created as a Classic Load Balancer
in AWS. Classic LBs allow you to assign a granular security group for
the traffic going through that LB. We have now switched the
documentation to create a Network Load Balancer instead. Network LBs do
not terminate connections and they preserve the original requestor's IP
address. This means the security group of the instance it is routed to
must allow traffic from that IP. Thus, it is necessary to add port 2222
to the general BOSH-deployed VMs security group; there is not a way to
target just the compute / Diego Brain instances that handle SSH routing.

Signed-off-by: Brian Upton <bupton@vmware.com>