azure: update instructions for Azure CLI changes
I tried to run the instructions to create an Azure service principle, but found that the Azure CLI had changed (I'm using 2.44.1). I worked out some alternative instructions that work for me and are in this PR. I'm not an Azure expert, so perhaps an SME should take a look over this first.